🎉🎊🍾 [New Feature] (code + test + script + config) Finish election feature of SmoothCrawler-Cluster and Add integration testing in CI.

+++ test/integration_test/_utils/zookeeper.py
@@ -1,10 +1,11 @@
 from smoothcrawler_cluster._utils.zookeeper import (
     _BaseZookeeperNode,
+    ZookeeperRecipe,
     _BaseZookeeperClient, ZookeeperClient,
     _BaseZookeeperListener
 )
 from kazoo.client import KazooClient
-from kazoo.exceptions import NodeExistsError
+from kazoo.exceptions import NodeExistsError, NoNodeError
 from typing import Type, TypeVar, Generic
 import pytest
 import random
@@ -22,36 +23,35 @@ class TestZookeeperClient:
 
     __PyTest_ZK_Client: KazooClient = None
 
-    @pytest.fixture(scope="function")
+    @pytest.fixture(scope="function", autouse=True)
     def zk_cli(self) -> Generic[_BaseZookeeperClientType]:
         self.__PyTest_ZK_Client = KazooClient(hosts=Zookeeper_Hosts)
         self.__PyTest_ZK_Client.start()
 
+        self._remove_zk_node_path()
+
         return ZookeeperClient(hosts=Zookeeper_Hosts)
 
-    @staticmethod
-    def _remove_path_finally(test_item):
+    def _remove_zk_node_path(self):
+        if self.__PyTest_ZK_Client.exists(path=Test_Zookeeper_Path) is not None:
+            # Remove the metadata of target path in Zookeeper
+            self.__PyTest_ZK_Client.delete(path=Test_Zookeeper_Path)
 
-        def _(self, zk_cli: Generic[_BaseZookeeperClientType]):
-            try:
-                test_item(self, zk_cli)
-            finally:
-                # Remove the metadata of target path in Zookeeper
-                self.__PyTest_ZK_Client.delete(path=Test_Zookeeper_Path)
-
-        return _
-
-    @_remove_path_finally
     def test_exist_path_with_exist_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
         # Test with a path which already exists ---> function should return True.
         zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
         assert zk_cli.exist_node(path=Test_Zookeeper_Path) is not None, "It should exist the path (node) it created."
 
+    def test_restrict_exist_path_with_exist_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
+        # Test with a path which already exists ---> function should return True.
+        zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
+        assert zk_cli.restrict_exist_node(path=Test_Zookeeper_Path, restrict=ZookeeperRecipe.ReadLock, identifier="pytest") is not None, \
+            "It should exist the path (node) it created."
+
     def test_exist_path_with_not_exist_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
         # Test with a path which doesn't exist ---> function should return False.
         assert zk_cli.exist_node(path=Test_Zookeeper_Not_Exist_Path) is None, "It should NOT exist the path (node)."
 
-    @_remove_path_finally
     def test_get_path_with_exist_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
         # Test with a path which already exists ---> function should return a object which is _BaseZookeeperClientType type.
         _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
@@ -61,13 +61,24 @@ class TestZookeeperClient:
         assert _zk_path.path == Test_Zookeeper_Path, f"The path of *ZookeeperPath* should be equal to {Test_Zookeeper_Path}."
         assert _zk_path.value == Test_Zookeeper_String_Value, f"The value of *ZookeeperPath* should be equal to \"{Test_Zookeeper_String_Value}\"."
 
+    def test_restrict_get_path_with_exist_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
+        # Test with a path which already exists ---> function should return a object which is _BaseZookeeperClientType type.
+        _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
+        assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
+
+        _zk_path = zk_cli.restrict_get_node(path=Test_Zookeeper_Path, restrict=ZookeeperRecipe.ReadLock, identifier="pytest")
+        assert _zk_path.path == Test_Zookeeper_Path, f"The path of *ZookeeperPath* should be equal to {Test_Zookeeper_Path}."
+        assert _zk_path.value == Test_Zookeeper_String_Value, f"The value of *ZookeeperPath* should be equal to \"{Test_Zookeeper_String_Value}\"."
+
     def test_get_path_with_not_exist_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
         # Test with a path which doesn't exist ---> function should
-        _zk_path = zk_cli.get_node(path=Test_Zookeeper_Not_Exist_Path)
-        assert _zk_path.path is Test_Zookeeper_Not_Exist_Path, f"The path of *ZookeeperPath* should be equal to {Test_Zookeeper_Not_Exist_Path}."
-        assert _zk_path.value is None, "The value of *ZookeeperPath* should be None because it doesn't exist recently."
+        try:
+            _zk_path = zk_cli.get_node(path=Test_Zookeeper_Not_Exist_Path)
+        except NoNodeError:
+            assert True, "Work finely."
+        else:
+            assert False, "It should raise an exception 'NoNodeError'."
 
-    @_remove_path_finally
     def test_create_path_without_value(self, zk_cli: Generic[_BaseZookeeperClientType]):
         _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path)
         assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
@@ -76,7 +87,6 @@ class TestZookeeperClient:
         assert _zk_path.path == Test_Zookeeper_Path, f"The path of *ZookeeperPath* should be equal to {Test_Zookeeper_Path}."
         assert _zk_path.value == "", "The value of *ZookeeperPath* should be equal to empty string because it created it without any value."
 
-    @_remove_path_finally
     def test_create_path_with_str_value(self, zk_cli: Generic[_BaseZookeeperClientType]):
         _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
         assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
@@ -85,7 +95,6 @@ class TestZookeeperClient:
         assert _zk_path.path == Test_Zookeeper_Path, f"The path of *ZookeeperPath* should be equal to {Test_Zookeeper_Path}."
         assert _zk_path.value == Test_Zookeeper_String_Value, f"The value of *ZookeeperPath* should be equal to {Test_Zookeeper_String_Value}."
 
-    @_remove_path_finally
     def test_create_path_with_bytes_value(self, zk_cli: Generic[_BaseZookeeperClientType]):
         _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_Bytes_Value)
         assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
@@ -94,7 +103,6 @@ class TestZookeeperClient:
         assert _zk_path.path == Test_Zookeeper_Path, f"The path of *ZookeeperPath* should be equal to {Test_Zookeeper_Path}."
         assert _zk_path.value == Test_Zookeeper_String_Value, f"The value of *ZookeeperPath* should be equal to {Test_Zookeeper_String_Value}."
 
-    @_remove_path_finally
     def test_create_path_with_already_exist_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
         _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_Bytes_Value)
         assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
@@ -122,7 +130,15 @@ class TestZookeeperClient:
             else:
                 assert False, "It should raise an exception about 'TypeError' because the data type of function argument *value* is invalid."
 
-    @_remove_path_finally
+    # def test_restrict_create_path_with_str_value(self, zk_cli: Generic[_BaseZookeeperClientType]):
+    #     _creating_result = zk_cli.restrict_create_node(
+    #         path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value, restrict=ZookeeperRecipe.WriteLock, identifier="pytest")
+    #     assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
+    #
+    #     _zk_path = zk_cli.get_node(path=Test_Zookeeper_Path)
+    #     assert _zk_path.path == Test_Zookeeper_Path, f"The path of *ZookeeperPath* should be equal to {Test_Zookeeper_Path}."
+    #     assert _zk_path.value == Test_Zookeeper_String_Value, f"The value of *ZookeeperPath* should be equal to {Test_Zookeeper_String_Value}."
+
     def test_get_value_from_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
         _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
         assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
@@ -130,13 +146,30 @@ class TestZookeeperClient:
         _value = zk_cli.get_value_from_node(path=Test_Zookeeper_Path)
         assert _value == Test_Zookeeper_String_Value, f"The value of *ZookeeperPath* should be equal to {Test_Zookeeper_String_Value}."
 
-    @_remove_path_finally
+    def test_restrict_get_value_from_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
+        _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
+        assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
+
+        _value = zk_cli.restrict_get_value_from_node(path=Test_Zookeeper_Path, restrict=ZookeeperRecipe.ReadLock, identifier="pytest")
+        assert _value == Test_Zookeeper_String_Value, f"The value of *ZookeeperPath* should be equal to {Test_Zookeeper_String_Value}."
+
     def test_set_value_to_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
         _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
         assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
 
         _new_value = "new zookeeper value"
         zk_cli.set_value_to_node(path=Test_Zookeeper_Path, value=_new_value)
+        assert zk_cli.get_node(path=Test_Zookeeper_Path).value == _new_value, \
+            f"The new value of *ZookeeperPath* should be equal to {_new_value} (by function *get_path*) because it has been modified."
+        assert zk_cli.get_value_from_node(path=Test_Zookeeper_Path) == _new_value, \
+            f"The new value of *ZookeeperPath* should be equal to {_new_value} (by function *get_value_from_path*) because it has been modified."
+
+    def test_restrict_set_value_to_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
+        _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
+        assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
+
+        _new_value = "new zookeeper value"
+        zk_cli.restrict_set_value_to_node(path=Test_Zookeeper_Path, value=_new_value, identifier="pytest")
         assert zk_cli.get_node(path=Test_Zookeeper_Path).value == _new_value, \
             f"The new value of *ZookeeperPath* should be equal to {_new_value} (by function *get_path*) because it has been modified."
         assert zk_cli.get_value_from_node(path=Test_Zookeeper_Path) == _new_value, \
@@ -149,11 +182,15 @@ class TestZookeeperClient:
         assert _zk_path.path == Test_Zookeeper_Path, f"The path of *ZookeeperPath* should be equal to {Test_Zookeeper_Path}."
         assert _zk_path.value == Test_Zookeeper_String_Value, f"The value of *ZookeeperPath* should be equal to {Test_Zookeeper_String_Value}."
 
-        _creating_result = zk_cli.remove_node(path=Test_Zookeeper_Path)
+        _creating_result = zk_cli.delete_node(path=Test_Zookeeper_Path)
 
-        _zk_path = zk_cli.get_node(path=Test_Zookeeper_Path)
-        assert _zk_path.path == Test_Zookeeper_Path, f"The path of *ZookeeperPath* should be equal to {Test_Zookeeper_Path}."
-        assert _zk_path.value is None, "The value of *ZookeeperPath* should be None because the path be deleted."
+        try:
+            _zk_path = zk_cli.get_node(path=Test_Zookeeper_Path)
+        except NoNodeError:
+            assert True, "Work finely."
+        else:
+            assert False, "It should raise an exception 'NoNodeError'."
+
 
 
 class TestZookeeperListener:
